Drop use of eval-when to bootstrap uptimes saving See https://github.com/melpa/melpa/pull/4647#issuecomment-289174171
Not only will it cause cl/cl-lib issues, I imagine, it doesn't really
seem to serve any useful purpose. I have a fuzzy recollection that it
did at the time I first wrote this but I'll be buggered if I can
remember what it was about. So... progn it is.