:zap: Don't save if we're not an interactive Emacs
I can't see a good reason why I'd ever want to track the uptimes of an Emacs
session that isn't being used by me. I also think this might be the reason
why, over time, I've been getting stray copies of the uptimes file in my
~/.emacs.d directory.

The way I have things working, I tell uptimes to keep the tracking file in
~/.local/share/emacs. But now and again I get a rogue copy will appear back
in ~/.emacs.d. Originally I thought the cause of this was the save on
startup, so f4727bd4f3421b905c5a0cc586b0d7c72389c046 was made to try and
tackle that. However, I've still being getting the issue.

I think it may be down to the fact that emacs gets called as a compiler when
I'm installing and upgrading stuff.

So, I don't know for sure this is the fix I'm looking for, but it can't
hurt.